requirements.txt
absl-py==1.0.0
bidict==0.21.4
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-SocketIO==5.1.1
flatbuffers==1.12
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
myModule==1.0.0
mypy-extensions==0.4.3
numpy==1.22.2
opencv-python==4.5.3.56
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.5.1
pybind11==2.9.1
pyFirmata==1.1.0
pyserial==3.5
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
RPi.GPIO==0.7.1
six==1.16.0
tflite-runtime==2.7.0
tflite-support==0.3.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.3
